dom/tests/mochitest/bugs/test_bug739038.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=739038
-->
<head>
  <title>Test for Bug 739038</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=739038">Mozilla Bug 739038</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 739038 **/

SimpleTest.waitForExplicitFinish();

SpecialPowers.pushPrefEnv({"set": [["dom.disable_window_print", true]]}, function() {
  var caughtException = false;
  try {
    window.print();
  } catch (e) {
    caughtException = true;
  }
  ok(caughtException, "window.print should throw an exception");
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
